# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title; generated import paths depend on it.
  title: Api
  version: 0.2.0
  description: LootBot multi-tenant store management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [catalog]
      summary: List categories for a store
      security:
        - sessionCookie: []
      parameters:
        - name: storeId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Categories and pagination details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /stores/{storeId}/categories:
    post:
      operationId: createCategory
      tags: [catalog]
      summary: Add a category to a store
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/StoreId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /categories/{categoryId}:
    parameters:
      - $ref: "#/components/parameters/CategoryId"
    patch:
      operationId: updateCategory
      tags: [catalog]
      summary: Update an owned category
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCategory
      tags: [catalog]
      summary: Soft-delete an owned category
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Category archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /products:
    get:
      operationId: listProducts
      tags: [catalog]
      summary: Search and paginate store products
      security:
        - sessionCookie: []
      parameters:
        - name: storeId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
        - name: categoryId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Products and pagination details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /stores/{storeId}/products:
    post:
      operationId: createProduct
      tags: [catalog]
      summary: Add a product to a store
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/StoreId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /products/{productId}:
    parameters:
      - $ref: "#/components/parameters/ProductId"
    patch:
      operationId: updateProduct
      tags: [catalog]
      summary: Update an owned product
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProduct
      tags: [catalog]
      summary: Soft-delete an owned product
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Product archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: lootbot_session
  parameters:
    StoreId:
      name: storeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CategoryId:
      name: categoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        enum: [10, 25, 50, 100]
        default: 25
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Sign in required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    CategoryInput:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        description:
          type: string
          maxLength: 500
      required: [name]
    CategoryUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
        description:
          type: string
          maxLength: 500
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: ["string", "null"]
        productCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, storeId, name, description, productCount, createdAt]
    CategoryPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      required: [items, page, pageSize, total]
    ProductInput:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 5000
        price:
          type: number
          minimum: 0
          maximum: 999999999
        stock:
          type: integer
          minimum: 0
          maximum: 100000000
        categoryId:
          type: ["string", "null"]
          format: uuid
        sku:
          type: ["string", "null"]
          maxLength: 80
        imageUrl:
          type: ["string", "null"]
          format: uri
          maxLength: 2000
        isPublished:
          type: boolean
      required: [name, description, price, stock, isPublished]
    ProductUpdate:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 160
        description:
          type: string
          maxLength: 5000
        price:
          type: number
          minimum: 0
          maximum: 999999999
        stock:
          type: integer
          minimum: 0
          maximum: 100000000
        categoryId:
          type: ["string", "null"]
          format: uuid
        sku:
          type: ["string", "null"]
          maxLength: 80
        imageUrl:
          type: ["string", "null"]
          format: uri
          maxLength: 2000
        isPublished:
          type: boolean
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        categoryId:
          type: ["string", "null"]
          format: uuid
        categoryName:
          type: ["string", "null"]
        sku:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
        imageUrl:
          type: ["string", "null"]
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, storeId, categoryId, categoryName, sku, name, description, price, stock, imageUrl, isPublished, createdAt]
    ProductPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      required: [items, page, pageSize, total]
...