# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LootBot multi-tenant store management API
 * OpenAPI spec version: 0.2.0
 */

export interface CategoryUpdate {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name?: string;
  /** @maxLength 500 */
  description?: string;
}