# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LootBot multi-tenant store management API
 * OpenAPI spec version: 0.2.0
 */

export interface Product {
  id: string;
  storeId: string;
  /** @nullable */
  categoryId: string | null;
  /** @nullable */
  categoryName: string | null;
  /** @nullable */
  sku: string | null;
  name: string;
  description: string;
  price: number;
  stock: number;
  /** @nullable */
  imageUrl: string | null;
  isPublished: boolean;
  createdAt: Date;
}