# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LootBot multi-tenant store management API
 * OpenAPI spec version: 0.2.0
 */

export interface ProductInput {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /** @maxLength 5000 */
  description: string;
  /**
     * @minimum 0
     * @maximum 999999999
     */
  price: number;
  /**
     * @minimum 0
     * @maximum 100000000
     */
  stock: number;
  /** @nullable */
  categoryId?: string | null;
  /**
     * @maxLength 80
     * @nullable
     */
  sku?: string | null;
  /**
     * @maxLength 2000
     * @nullable
     */
  imageUrl?: string | null;
  isPublished: boolean;
}